+++ package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "@hookform/resolvers": "^3.9.0",
     "axios": "^1.7.7",
+    "dayjs": "^1.11.13",
     "primereact": "^10.8.4",
     "quill": "^2.0.2",
     "react": "^18.3.1",
