+++ src/utils/enums/index.ts
@@ -1,1 +1,1 @@
-export * from './OrderEnum'
+export * from './orderEnum'
