+++ src/utils/helpers/index.ts
@@ -1,4 +1,4 @@
 export * from './uploadToCloudinary'
 export * from './convertStringDate'
-export * from './Checkout'
+export * from './checkout'
 export * from './common'
